Fix https://issues.apache.org/bugzilla/show_bug.cgi?id=51162 Prevent possible NPE when removing a web application
git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc6.0.x/trunk@1142206 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/startup/HostConfig.java
@@ -1214,11 +1214,13 @@ public class HostConfig
                     log.warn(sm.getString
                              ("hostConfig.context.remove", app.name), t);
                 }
-                try {
-                    context.destroy();
-                } catch (Throwable t) {
-                    log.warn(sm.getString
-                             ("hostConfig.context.destroy", app.name), t);
+                if (context != null) {
+                    try {
+                        context.destroy();
+                    } catch (Throwable t) {
+                        log.warn(sm.getString
+                                ("hostConfig.context.destroy", app.name), t);
+                    }
                 }
                 // Delete all redeploy resources
                 for (int j = i + 1; j < resources.length; j++) {
